LayoutTests/fast/viewport/viewport-53.html

<head>
    <title>Check that unrecognized keyword translate to 0 for height.</title>
    <meta name="viewport" content="height=y20">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
